move culture input
