+++ src/backoffice/settings/cultures/components/index.ts
@@ -1,0 +1,1 @@
+import './input-culture-select/input-culture-select.element';
