+++ src/backoffice/settings/cultures/components/input-culture-select/input-culture-select.element.ts
@@ -5,15 +5,13 @@ import { FormControlMixin } from '@umbraco-ui/uui-base/lib/mixins';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { repeat } from 'lit/directives/repeat.js';
 import { UUIComboboxElement, UUIComboboxEvent } from '@umbraco-ui/uui';
-import { UmbCultureRepository } from '../../../settings/cultures/repository/culture.repository';
+import { UmbCultureRepository } from '../../repository/culture.repository';
 import { UmbChangeEvent } from '@umbraco-cms/backoffice/events';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
 import { CultureReponseModel } from '@umbraco-cms/backoffice/backend-api';
 
 @customElement('umb-input-culture-select')
 export class UmbInputCultureSelectElement extends FormControlMixin(UmbLitElement) {
-	
-
 	/**
 	 * Disables the input
 	 * @type {boolean}
@@ -104,7 +102,7 @@ export class UmbInputCultureSelectElement extends FormControlMixin(UmbLitElement
 				  `}
 		`;
 	}
-	
+
 	static styles = [UUITextStyles, css``];
 }
