+++ src/backoffice/settings/index.ts
@@ -15,6 +15,8 @@ import {
 import { UmbContextProviderController } from '@umbraco-cms/backoffice/context-api';
 import type { UmbEntrypointOnInit } from '@umbraco-cms/backoffice/extensions-api';
 
+import './cultures/components';
+
 export const manifests = [
 	...settingsSectionManifests,
 	...settingsMenuManifests,
